backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.13",
    "google-genai>=1.22.0",
    "google-cloud-storage>=2.10.0",
    "anthropic>=0.40.0",
    "uvicorn>=0.34.3",
    "google-cloud-aiplatform>=1.111.0",
    "pillow>=10.0.0",
    "python-jose[cryptography]>=3.5.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "pydantic-settings>=2.11.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88
select = ["E", "F", "I", "N", "W", "B", "C4", "UP", "ARG", "SIM", "TCH", "TID", "Q"]
ignore = ["E501", "B008", "F401", "F841", "W293", "ARG001", "N815"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
